Worked on user registration by Company Admin and Normal User

+++ infrastructure/src/main/java/org/arispay/mappers/UserMapper.java
@@ -49,7 +49,7 @@ public abstract class UserMapper {
 		List<UserCompany> userCompanies = new ArrayList<>();
 		if(!(userCompanyDtos == null)) {
 			for(UserCompanyDto userCompanyDto : userCompanyDtos) {
-				Company company = companyRepository.findById(Long.valueOf(userCompanyDto.getCompanyId())).orElse(null);
+				Company company = companyRepository.findById(userCompanyDto.getCompanyId()).orElse(null);
 				UserCompany userCompany =  new UserCompany(company, userCompanyDto.isDefault());
 				userCompany.setId(userCompanyDto.getId());
 				userCompanies.add(userCompany);
@@ -66,7 +66,7 @@ public abstract class UserMapper {
 
 		List<UserCompanyDto> userCompanies = new ArrayList<>();
 		for( UserCompany company: companies ) {
-			userCompanies.add(new UserCompanyDto(company.getId(),company.getCompany().getId().intValue(), company.isDefault()));
+			userCompanies.add(new UserCompanyDto(company.getId(),company.getCompany().getId(), company.isDefault()));
 		}
 
 		return userCompanies;
